Moving to kubernetes package
